Adds a bare-bones setup.py.
It doesn't *do* anything right now, but we'll use it as the cornerstone
of the environment we'll build on top of it.

When specifying dependencies, I like to use explicit version numbers.
Upgrading direct dependencies should be an explicit operation. You certainly
don't want, for example, Django 1.5 to be installed if your application was
built against 1.4 and has not been tested against 1.5.

Arguments to setup():

- install_requires and version should be clear. These are the dependencies
  and version string, respectively, of your package.
- name is your, er, package name. Unless you know what you're doing, it's
  best to just name this after your top-level package. ("takestock", in this
  case.)
- package_dir and packages are magical incantations that tell setuptools
  that src/ contains the code for our package.