Adds a skeleton makefile.
The makefile will be used to manage the local dev environment. Usually
you see make in C-based projects, but it's actually very useful for
Python projects as well. (As I hope you'll soon see.)

The basic make syntax is:

target :
<tab>commands for target

When you run `make target`, it runs the commands defined for that
target. (Well, not exactly, but we'll get to the exact rules as we
go along.)

So here, we've defined a single target, `make help`, which prints a
help message for our makefile. There are four things to note here:

* Per PEP8 (the canonical Python style guide) Python code should use
  spaces, not tabs. Tabs-vs-spaces for Javascript, CSS, HTML, etc. are
  arguable (and I'm going to argue for spaces). But tabs are
  non-negotiable for Makefiles. If you use spaces, things will not work.
* By default, make will print each command prior to running it. When we're
  echo-ing the help, we don't want make to do this. @ suppresses the
  behavior. In general, you want to use @ for echo commands but nothing
  else.
* If you don't supply a target, make uses the first one defined. So, in
  our case, running `make` will print the help.
* There are two versions of make, "BSD make" and "GNU make". When looking
  at documentation, you want BSD. BSD make is what's on your system.