Adds initial `make env` target.
This is the first step on making the environment easy to set up. The
end goal is to get the developer setup instructions down to:

    git clone ...
    cd takestock
    make env

Obviously, we need a place to install our code and dependencies. We'll
do so in a virtualenv in the directory .virtualenvs/development/.

Why not ~/.virtualenvs/ as encouraged by virtualenvwrapper? Because we
want to keep the entire dev setup underneath the top-level takestock/
directory. That way, when you want to get rid of takestock, you simply
delete the directory.

Why not takestock/.virtualenv/? I can imagine a situation where it might
be useful to have a separate "testing" virtualenv. That is, you'd create
an sdist, install it into a fresh environment, then run the unit tests to
make sure you aren't forgetting loose ends in your packaging process. Where
should that testing environment live? .virtualenvs/testing/ of course.

For help on the specific arguments to pip, run `pip help install`. It's
pretty straightforward. We're telling it to run `setup.py develop` on the
takestock package and install the stuff in requirements.txt, all into the
.virtualenvs/development/.

This commit also shows some new things about make.

When make processes a target, it first looks for a file with the same
path as the target. If the file is missing, make runs the commands for
the target, assuming that those commands will build the file. If the
file already exists, make *will not* run the commands again, assuming that
they were already run previously.

Targets that don't actually represent files are called "phony" targets.
Our help target is one of those; running `make help` will never produce a
file named 'help' in the current directory.

Our .virtualenvs/development/bin/python target, however, is not phony.
It represents a real file that will exist after we run `make env`. The first
time it's run, that file doesn't exist so make runs the pip command.
Subsequent runs will see that the .virtualenvs/development/bin/python file
already exists, and will skip the pip command. Perfect.

The other target, `env`, is a phony target, which we've already seen with
`help`. What's new is the information to the right of the colon. (And
perhaps that a target doesn't have to define any commands. And perhaps
the line continuation convention [backslash]. But I'm assuming the latter
two aren't too foreign.)

A target can depend on other targets. The dependencies are listed to the
right of the colon, space-separated. When you run `make env`, make first
processes all the dependencies, then moves on to the target.

Right now, `make env` is essentially an alias for
`make .virtualenvs/development/bin/python`, but in the future you might
use it add dependencies for other development subsystems. (Like vagrant
virtual machines, or non-Python dev tools such as LESS or SASS.)