Moves development virtualenv path into a variable.
Variables in make are a little weird compared to what you're used to
in other languages. And you should especially not think of them like
you do in the shell (which is tempting, and a trap I fell into early
on). Make is not really a programming language, no matter how tempting
it is to start wanting it to be one. (This might not make sense to you
right now, but keep it in mind if you decide to pick up make. My basic
rule of thumb is "if I'm trying to use an if-statement, then I'm doing
it wrong").

With that caveat out of the way, all we've done here is put the
development virtualenv directory into a variable so it's easier to change
in the future should you decide to rearrange or rename your virtualenv
structure.