Changes the relative path to an absolute path.
I don't think this is strictly necessary, but I like dealing in
absolute paths wherever possible.

This also shows one of the "special" variable forms provided by make. If
you need to run a command, use the form $(shell <the command>). Like
the caveat on variables, this is good for simple stuff like getting the pwd,
but if you find yourself shelling out a lot you're probably doing it wrong.