Allows for alternately-named pip commands.
On my system, I have pip installed for each different version of Python.
(Not by choice; this is forced on me by MacPorts.) Each one is named
pip-x.y, where x.y is the Python version. So we need a way for me to tell
make to use `pip-2.7` when it's running on my machine, but just `pip`
otherwise.

To do so, we use the $(PIP) variable. The ?= syntax means "set to the
value on the right, unless the user has explicitly provided a value". To
provide a value, I run `make env` like this:

    make PIP=pip-2.7 env

Annoying, but only for developers that have annoying setups with non-standard
command names.