Moves settings into the takestock package.
Over the next few commits, we're going to flesh out the settings
structure. I used to think it was a bad idea to have your settings
in your app. And it is, if you intend to open-source and distribute
your app.

But for private projects, it makes a lot of sense to put as much in
the package as possible. There is practically no case where you're going
to want to re-use these apps, much less with different templates or
wildly different settings.