First crack at distilling settings into development/production.
Each of the modified files has a docstring at the top with a rough
explanation of what's in the file.

For production settings, I haven't changed much. Development settings
are basically all new, though. Everything should be pretty straightforward.

Note that this is still untested, so I'm not 100% sure it works. Getting
`syncdb` up and running is the next order of business.