Adds the takestock app to INSTALLED_APPS.
Everything appears to work! My plan from here is to clean up some loose
ends and try to nominally figure out the deployment situation. Then, if I
have the energy to do so, I'll take a crack at refactoring parts of the app
into the way I would approach it. (I'll do that in a separate branch. If you
don't like it you can toss it away.)